+++ libs/platform/src/lib/table/table.component.ts
@@ -18,6 +18,7 @@ import {
     Output,
     QueryList,
     SimpleChanges,
+    TrackByFunction,
     ViewChild,
     ViewChildren,
     ViewEncapsulation
@@ -26,7 +27,7 @@ import { DOCUMENT } from '@angular/common';
 import { BehaviorSubject, isObservable, merge, Observable, of, Subscription } from 'rxjs';
 import { debounceTime, distinctUntilChanged, filter, map, startWith, switchMap } from 'rxjs/operators';
 
-import { ContentDensityEnum, FdDropEvent, RtlService } from '@fundamental-ngx/core/utils';
+import { ContentDensityEnum, ContentDensityService, FdDropEvent, RtlService } from '@fundamental-ngx/core/utils';
 import { TableRowDirective } from '@fundamental-ngx/core/table';
 
 import { getNestedValue, isDataSource, isString } from '@fundamental-ngx/platform/shared';
@@ -58,7 +59,8 @@ import {
     TableRowSelectionChangeEvent,
     TableRowsRearrangeEvent,
     TableRowToggleOpenStateEvent,
-    TableSortChangeEvent
+    TableSortChangeEvent,
+    RowComparator
 } from './models';
 import { TableColumnResizeService } from './table-column-resize.service';
 import { TableColumnResizableSide } from './directives/table-cell-resizable.directive';
@@ -126,7 +128,8 @@ let tableUniqueId = 0;
         '[class.fd-table--condensed]': 'contentDensity === CONTENT_DENSITY.CONDENSED',
         '[class.fd-table--no-horizontal-borders]': 'noHorizontalBorders || noBorders',
         '[class.fd-table--no-vertical-borders]': 'noVerticalBorders || noBorders',
-        '[class.fdp-table--tree]': '_rowsDraggable'
+        '[class.fd-table--tree]': 'isTreeTable',
+        '[class.fd-table--group]': '_isGroupTable'
     }
 })
 export class TableComponent<T = any> extends Table implements AfterViewInit, OnDestroy, OnChanges, OnInit {
@@ -253,6 +256,21 @@ export class TableComponent<T = any> extends Table implements AfterViewInit, OnD
     /** Value with the key of the row item's field to compute semantic state of the row.  */
     @Input()
     semanticHighlighting: string;
+
+    /**
+     * Tracking function that will be used to check the differences in data changes. 
+     * Used similarly to `ngFor` `trackBy` function. 
+     * Accepts a function that takes two parameters, index and item.
+     */
+    @Input()
+    trackBy: TrackByFunction<T>;
+
+    /** 
+     * An optional function, that identifies uniqueness of a particular row.
+     * Table component uses it to be able to preserve selection when data list is changed.
+     */
+    @Input()
+    rowComparator: RowComparator<T>;
 
     /** Event fired when table selection has changed. */
     @Output()
@@ -439,6 +457,16 @@ export class TableComponent<T = any> extends Table implements AfterViewInit, OnD
     /** @hidden */
     _scrollBarWidth = 0;
 
+    /** 
+     * @hidden
+     * Mappping function for the trackBy, provided by the user. 
+     * Is needed, because we are wrapping user supplied data into a `TableRow` class.
+     */
+    _rowTrackBy: TrackByFunction<TableRow<T>>;
+
+    /** @hidden */
+    _isGroupTable = false;
+
     /** @hidden */
     get _isShownSelectionColumn(): boolean {
         return this.selectionMode !== SelectionMode.NONE;
@@ -486,6 +514,9 @@ export class TableComponent<T = any> extends Table implements AfterViewInit, OnD
     /** @hidden */
     private _dragDropInProgress = false;
 
+    /** @hidden Is used to identify whether the `contentDensity` property was set by the user manually. */
+    private contentDensityManuallySet = false;
+
     /** @hidden */
     private get _selectionColumnWidth(): number {
         return this._isShownSelectionColumn ? SELECTION_COLUMN_WIDTH.get(this.contentDensity) : 0;
@@ -504,15 +535,25 @@ export class TableComponent<T = any> extends Table implements AfterViewInit, OnD
         private readonly _tableScrollDispatcher: TableScrollDispatcherService,
         private readonly _tableColumnResizeService: TableColumnResizeService,
         @Inject(DOCUMENT) private readonly _document: Document | null,
-        @Optional() private readonly _rtlService: RtlService
+        @Optional() private readonly _rtlService: RtlService,
+        @Optional() private readonly _contentDensityService: ContentDensityService,
     ) {
         super();
+        this._trackContentDensityChanges();
     }
 
     /** @hidden */
     ngOnChanges(changes: SimpleChanges): void {
         if ('loading' in changes) {
             this._tableService.setTableLoading(this.loading);
+        }
+
+        if ('trackBy' in changes) {
+            this._rowTrackBy = typeof this.trackBy === 'function' ? (index, item) => this.trackBy(index, item.value) : undefined;
+        }
+
+        if (changes.contentDensity?.currentValue) {
+            this.contentDensityManuallySet = true;
         }
 
         // changes below should be checked only after view is initialized
@@ -538,6 +579,8 @@ export class TableComponent<T = any> extends Table implements AfterViewInit, OnD
     /** @hidden */
     ngOnInit(): void {
         this._calculateScrollbarWidth();
+
+        this._isGroupTable = this.initialGroupBy?.length > 0;
     }
 
     /** @hidden */
@@ -1057,6 +1100,11 @@ export class TableComponent<T = any> extends Table implements AfterViewInit, OnD
     }
 
     /** @hidden */
+    _columnTrackBy(index: number, column: TableColumn): string {
+        return column.name;
+    }
+
+    /** @hidden */
     private _isDroppedInsideItself(dropRow: TableRow, dragRow: TableRow): boolean {
         const dropRowParents = this._getRowParents(dropRow);
         return !!dropRowParents.find(row => row === dragRow);
@@ -1145,10 +1193,9 @@ export class TableComponent<T = any> extends Table implements AfterViewInit, OnD
 
         this._subscriptions.add(
             this._dataSourceItemsSubject
-                .asObservable()
                 .pipe(
                     // map source items to table rows
-                    switchMap((source: T[]) => of(this._createTableRowsByDataSourceItems(source))),
+                    map((source: T[]) => this._createTableRowsByDataSourceItems(source)),
                     // Insert items to show groups
                     switchMap((rows: TableRow[]) =>
                         this.isTreeTable
@@ -1227,6 +1274,7 @@ export class TableComponent<T = any> extends Table implements AfterViewInit, OnD
         this._subscriptions.add(
             this._tableService.columnsChange.subscribe((event: ColumnsChange) => {
                 this._calculateVisibleColumns();
+                this.recalculateTableColumnWidth();
 
                 this.columnsChange.emit(new TableColumnsChangeEvent(this, event.current, event.previous));
             })
@@ -1253,9 +1301,11 @@ export class TableComponent<T = any> extends Table implements AfterViewInit, OnD
             return this._createTreeTableRowsByDataSourceItems(source);
         }
 
+        const selectedRowsMap = this._getSelectionStatusByRowValue(source);
+
         return source
             .map((item: T, index: number) => {
-                const row = new TableRow(TableRowType.ITEM, false, index, item);
+                const row = new TableRow(TableRowType.ITEM, !!selectedRowsMap.get(item), index, item);
                 row.navigatable = this._isRowNavigatable(item, this.rowNavigatable);
 
                 return row;
@@ -1266,11 +1316,13 @@ export class TableComponent<T = any> extends Table implements AfterViewInit, OnD
     private _createTreeTableRowsByDataSourceItems(source: T[]): TableRow<T>[] {
         const rows: TableRow<T>[] = [];
 
+        const selectedRowsMap = this._getSelectionStatusByRowValue(source);
+
         source.forEach((item: T, index: number) => {
             const hasChildren = item.hasOwnProperty(this.relationKey)
                 && Array.isArray(item[this.relationKey])
                 && item[this.relationKey].length;
-            const row = new TableRow(hasChildren ? TableRowType.TREE : TableRowType.ITEM, false, index, item);
+            const row = new TableRow(hasChildren ? TableRowType.TREE : TableRowType.ITEM, !!selectedRowsMap.get(item), index, item);
 
             row.expanded = false;
             row.navigatable = this._isRowNavigatable(item, this.rowNavigatable);
@@ -1290,6 +1342,30 @@ export class TableComponent<T = any> extends Table implements AfterViewInit, OnD
         });
 
         return rows;
+    }
+
+    /** 
+     * @hidden
+     * Runs `rowComparator` function against checked rows and compares them with the new `source`
+     * If matched, creates an association between the source item and checked status of corresponding row.
+     * 
+     * @returns `Map` object with the `checked` status for particular source item
+     */
+    private _getSelectionStatusByRowValue(source: T[]): Map<T, boolean> {
+        const rowMap = new Map<T, boolean>();
+        if (
+            (this.selectionMode === SelectionMode.SINGLE || this.selectionMode === SelectionMode.MULTIPLE) &&
+            typeof this.rowComparator === 'function'
+        ) {
+            const checkedRows = this._tableRows.filter((r) => r.checked);
+            checkedRows.forEach((row) => {
+                const found = source.find((e) => this.rowComparator(row.value, e));
+                if (found) {
+                    rowMap.set(found, row.checked);
+                }
+            });
+        }
+        return rowMap;
     }
 
     /** @hidden */
@@ -1398,6 +1474,7 @@ export class TableComponent<T = any> extends Table implements AfterViewInit, OnD
     private _buildGroupRulesMap(state = this.getTableState()): void {
         const groupMap = new Map(state.groupBy.map((rule) => [rule.field, rule]));
         this._groupRulesMapSubject.next(groupMap);
+        this._isGroupTable = groupMap.size > 0;
     }
 
     /** @hidden */
@@ -1802,5 +1879,16 @@ export class TableComponent<T = any> extends Table implements AfterViewInit, OnD
         }
 
         return !!rowNavigatable;
+    }
+
+    private _trackContentDensityChanges(): void {
+        if (this._contentDensityService) {
+            this._subscriptions.add(this._contentDensityService._contentDensityListener
+                .pipe(filter(() => !this.contentDensityManuallySet))
+                .subscribe((density) => {
+                    this.contentDensity = density as ContentDensityEnum;
+                    this._cdr.markForCheck();
+                }))
+        }
     }
 }